+++ packages/guide/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### TBD
+
+- Added concerned monkey shop
+
 ### 0.11.0 2023-09-07
 
 #### Added